devfreq: Fix type of frequency table
Frequencies are always expressed as ulong, but devfreq's table is of
type uint *. Fix the mismatch.

Reviewed-on: http://git-master/r/1123438
Tested-by: Shridhar Rasal <srasal@nvidia.com>

+++ include/linux/devfreq.h
@@ -105,7 +105,7 @@ struct devfreq_dev_profile {
 	int (*set_low_wmark)(struct device *dev, unsigned int val);
 	void (*exit)(struct device *dev);
 
-	unsigned int *freq_table;
+	unsigned long *freq_table;
 	unsigned int max_state;
 };
